# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: APS - Automotive Platform System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /shops/{shopId}/bays:
    post:
      operationId: createShopBay
      tags: [bays]
      summary: Add a bay to a shop (owner only)
      parameters:
        - name: shopId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBayBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bay"
    get:
      operationId: listShopBays
      tags: [bays]
      summary: List bays for a shop
      parameters:
        - name: shopId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Bay"
  /bays/available:
    get:
      operationId: listAvailableBays
      tags: [bays]
      summary: Search bookable bays (active mechanic only)
      parameters:
        - name: jobCategory
          in: query
          required: false
          schema:
            type: string
            enum: [repair, diagnostic, maintenance, detailing]
        - name: minTier
          in: query
          required: false
          schema:
            type: string
            enum: [detailer, technician, senior, advanced, master]
        - name: jobId
          in: query
          required: false
          schema: { type: integer, minimum: 1 }
        - name: startsAt
          in: query
          required: false
          schema: { type: string, format: date-time }
        - name: durationHours
          in: query
          required: false
          schema: { type: number, exclusiveMinimum: 0, maximum: 24 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BayWithShop"
  /bays/{bayId}:
    get:
      operationId: getBay
      tags: [bays]
      summary: Get a single bay with its shop
      parameters:
        - name: bayId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BayWithShop"
    patch:
      operationId: updateBay
      tags: [bays]
      summary: Update a bay (owner only)
      parameters:
        - name: bayId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBayBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bay"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Shop:
      type: object
      properties:
        id:
          type: integer
        ownerId:
          type: integer
        organizationId:
          type: ["integer", "null"]
        partnerKind:
          type: string
          enum: [independent_shop, dealership, fleet, gsa]
          description: |
            Discriminates the partner type. `independent_shop` uses the classic
            bay-rental marketplace. `dealership`, `fleet`, and `gsa` (Government /
            GSA accounts) can additionally post overflow jobs at a reduced
            commission.
        name:
          type: string
        address:
          type: string
        city:
          type: string
        region:
          type: string
        zipCode:
          type: string
        lat:
          type: number
          nullable: true
        lng:
          type: number
          nullable: true
        phone:
          type: string
          nullable: true
        federalEin:
          type: string
          nullable: true
          description: Federal EIN / Tax ID — used to uniquely identify the partner business with the IRS.
        businessLicense:
          type: string
          nullable: true
          description: State or local business license number.
        insuranceCarrier:
          type: string
          nullable: true
          deprecated: true
          description: DEPRECATED. Per-vehicle insurance now lives on `Vehicle.insuranceCarrier`.
        insurancePolicyNumber:
          type: string
          nullable: true
          deprecated: true
          description: DEPRECATED. Per-vehicle insurance now lives on `Vehicle.insurancePolicyNumber`.
        status:
          type: string
          enum: [active, inactive]
        createdAt:
          type: string
          format: date-time
      required: [id, ownerId, organizationId, partnerKind, name, address, city, region, zipCode, status, createdAt]
...
    Bay:
      type: object
      properties:
        id:
          type: integer
        shopId:
          type: integer
        name:
          type: string
        hourlyRate:
          type: number
        equipment:
          type: array
          items:
            type: string
        allowedJobCategories:
          type: array
          items:
            type: string
            enum: [repair, diagnostic, maintenance, detailing]
        minMechanicTier:
          type: string
          enum: [detailer, technician, senior, advanced, master]
        autoApprove:
          type: boolean
        availabilityConfig:
          $ref: "#/components/schemas/BayAvailabilityConfig"
        status:
          type: string
          enum: [active, inactive]
        createdAt:
          type: string
          format: date-time
      required: [id, shopId, name, hourlyRate, equipment, allowedJobCategories, minMechanicTier, autoApprove, availabilityConfig, status, createdAt]
...
    BayWithShop:
      allOf:
        - $ref: "#/components/schemas/Bay"
        - type: object
          properties:
            shop:
              $ref: "#/components/schemas/Shop"
          required: [shop]
    CreateBayBody:
      type: object
      properties:
        name:
          type: string
        hourlyRate:
          type: number
          minimum: 0
        equipment:
          type: array
          items:
            type: string
        allowedJobCategories:
          type: array
          items:
            type: string
            enum: [repair, diagnostic, maintenance, detailing]
        minMechanicTier:
          type: string
          enum: [detailer, technician, senior, advanced, master]
        autoApprove:
          type: boolean
        availabilityConfig:
          $ref: "#/components/schemas/BayAvailabilityConfig"
      required: [name, hourlyRate, allowedJobCategories, minMechanicTier]
    UpdateBayBody:
      type: object
      properties:
        name:
          type: string
        hourlyRate:
          type: number
          minimum: 0
        equipment:
          type: array
          items:
            type: string
        allowedJobCategories:
          type: array
          items:
            type: string
            enum: [repair, diagnostic, maintenance, detailing]
        minMechanicTier:
          type: string
          enum: [detailer, technician, senior, advanced, master]
        autoApprove:
          type: boolean
        availabilityConfig:
          $ref: "#/components/schemas/BayAvailabilityConfig"
        status:
          type: string
          enum: [active, inactive]
    BayAvailabilityConfig:
      type: object
      additionalProperties: false
      properties:
        timezone:
          type: string
          enum: [UTC]
        weekly:
          type: array
          items:
            $ref: "#/components/schemas/BayAvailabilityWindow"
      required: [timezone, weekly]
    BayAvailabilityWindow:
      type: object
      additionalProperties: false
      properties:
        dayOfWeek:
          type: integer
          minimum: 0
          maximum: 6
        open:
          type: string
          pattern: "^([01][0-9]|2[0-3]):[0-5][0-9]$"
        close:
          type: string
          pattern: "^([01][0-9]|2[0-3]):[0-5][0-9]$"
      required: [dayOfWeek, open, close]
...